WebIDL/tests/submissions/heycam/DOMString-003.html

<!DOCTYPE html>
<title>DOMString</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.16 DOMString
//
// IDL DOMString values are represented with JS string values.

// feature: Attr.value (attribute) (of type DOMString)

test(function() {
  var e = document.createElement("span");
  e.setAttribute("data-test", "test");
  var a = e.getAttributeNode("data-test");
  assert_equals(typeof a.value, "string");
}, "Attr.value returns a value of the correct type");

// feature: Node.nodeName (attribute) (of type DOMString)

test(function() {
  assert_equals(typeof document.nodeName, "string");
}, "Node.nodeName returns a value of the correct type");
</script>
